requirements.txt
discord.py==1.2.3
git+git://github.com/kennethreitz/requests.git